# openapi-static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Perfect Schema API",
    "summary": "A deliberately precise API for contract testing.",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "description": "List users with optional filtering, sorting, and pagination.",
        "operationId": "listUsers",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by name or email",
              "title": "Q"
            },
            "description": "Filter by name or email"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^-?name$",
              "default": "name",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users.",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/UserCollection"
                },
                "example": {
                  "items": [],
                  "count": 0,
                  "links": {
                    "self": "/users"
                  }
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Total-Count": {
                "schema": {
                  "type": "integer"
                }
              },
              "Link": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create user",
        "description": "Create a new user.",
        "operationId": "createUser",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully.",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "User already exists.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "description": "RFC 7807 compatible problem details.",
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "examples": [
                        "https://example.com/problems/not-found"
                      ],
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "examples": [
                        "Resource not found"
                      ],
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "examples": [
                        404
                      ],
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "examples": [
                        "User with id ... was not found"
                      ],
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "examples": [
                        "/users/123"
                      ],
                      "title": "Instance",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{user_id}/orders": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List a user's orders",
        "description": "List orders for a single user.",
        "operationId": "listUserOrders",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/OrderCollection"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Total-Count": {
                "schema": {
                  "type": "integer"
                }
              },
              "Link": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "description": "RFC 7807 compatible problem details.",
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "examples": [
                        "https://example.com/problems/not-found"
                      ],
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "examples": [
                        "Resource not found"
                      ],
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "examples": [
                        404
                      ],
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "examples": [
                        "User with id ... was not found"
                      ],
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "examples": [
                        "/users/123"
                      ],
                      "title": "Instance",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OrderCollection": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderRead"
            },
            "type": "array",
            "title": "Items"
          },
          "count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Count"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          }
        },
        "type": "object",
        "required": [
          "count",
          "links"
        ],
        "title": "OrderCollection",
        "description": "Response for order list."
      },
...
      "OrderRead": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/OrderStatus",
            "default": "pending"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total",
            "examples": [
              "125.50"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "placed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Placed At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "total",
          "id",
          "user_id",
          "placed_at",
          "updated_at"
        ],
        "title": "OrderRead",
        "description": "Order representation."
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "shipped",
          "delivered",
          "cancelled"
        ],
        "title": "OrderStatus",
        "description": "Enumeration of the lifecycle states of an order."
      },
...
      "PaginationLinks": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self",
            "examples": [
              "/users?offset=0&limit=20"
            ]
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next",
            "examples": [
              "/users?offset=10&limit=5"
            ]
          },
          "prev": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev",
            "examples": [
              "/users?offset=0&limit=5"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "self"
        ],
        "title": "PaginationLinks",
        "description": "Link relations for paginated endpoints."
      },
...
      "UserCollection": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserRead"
            },
            "type": "array",
            "title": "Items"
          },
          "count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Count",
            "examples": [
              3
            ]
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          }
        },
        "type": "object",
        "required": [
          "count",
          "links"
        ],
        "title": "UserCollection",
        "description": "Response for user list endpoints."
      },
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "user@example.com"
            ]
          },
          "full_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name",
            "examples": [
              "Ada Lovelace"
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true,
            "examples": [
              true
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name"
        ],
        "title": "UserCreate",
        "description": "Schema for creating users."
      },
...
      "UserRead": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "user@example.com"
            ]
          },
          "full_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name",
            "examples": [
              "Ada Lovelace"
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true,
            "examples": [
              true
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time",
            "title": "Joined At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "id",
          "joined_at",
          "updated_at"
        ],
        "title": "UserRead",
        "description": "User representation in responses."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {
              "users:read": "Read user information",
              "users:write": "Modify users",
              "orders:read": "Read orders",
              "orders:write": "Modify orders"
            },
            "tokenUrl": "/token"
          }
        }
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}